refactor(twitch): reorganize API modules for better scalability
