+++ Desktop/Features/Integrations/Twitch/Services/TwitchApiModerationModule.cs
@@ -7,6 +7,32 @@ namespace StreamBoard.Features.Integrations.Twitch.Services
     public class TwitchApiModerationModule(TwitchAuthContext context, HttpClient http)
         : TwitchApiModule(context, http)
     {
+        public async Task DeleteChatMessages(string broadcasterId, string moderatorId, string? messageId = null)
+        {
+            var query = $"broadcaster_id={broadcasterId}&moderator_id={moderatorId}";
+
+            if (!string.IsNullOrWhiteSpace(messageId))
+            {
+                query += $"&message_id={messageId}";
+            }
+
+            try
+            {
+                await SendRequestInternal(HttpMethod.Delete, "/moderation/chat", query);
+            }
+            catch (Exception ex)
+            {
+                if (ex is Exceptions.TwitchApiException) throw;
+                throw new Exception($"Failed to delete chat messages: {ex.Message}", ex);
+            }
+        }
+
+        public async Task ClearChat(string broadcasterId, string moderatorId)
+            => await DeleteChatMessages(broadcasterId, moderatorId);
+
+        public async Task DeleteMessage(string broadcasterId, string moderatorId, string messageId)
+            => await DeleteChatMessages(broadcasterId, moderatorId, messageId);
+
         public async Task<TwitchShieldModeStatus?> GetShieldModeStatus(string broadcasterId, string moderatorId)
         {
             var query = $"broadcaster_id={broadcasterId}&moderator_id={moderatorId}";
